backend/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyodb==0.1.4
pyodbc==4.0.39
python-dotenv==1.0.0
Werkzeug==2.3.7
